refactor: 将chunkSize和totalSize的类型从Long改为Integer
为了统一数据类型并减少不必要的类型转换，将ChunkInfo类中的chunkSize和totalSize字段类型从Long改为Integer。同时，更新了相关的控制器和服务层代码以适配此变更。

+++ src/main/java/com/controller/MinioController.java
@@ -83,11 +83,13 @@ public class MinioController {
 
             // 其他可选参数
             if (params.containsKey("chunkSize")) {
-                chunkInfo.setChunkSize(Long.parseLong(params.get("chunkSize").toString()));
+                // 使用Integer.parseInt进行正确的字符串到整数的转换
+                chunkInfo.setChunkSize(Integer.parseInt(params.get("chunkSize").toString()));
             }
 
             if (params.containsKey("totalSize")) {
-                chunkInfo.setTotalSize(Long.parseLong(params.get("totalSize").toString()));
+                // 同样修复totalSize的转换
+                chunkInfo.setTotalSize(Integer.parseInt(params.get("totalSize").toString()));
             }
 
             if (params.containsKey("fileType")) {
@@ -116,8 +118,8 @@ public class MinioController {
     public R uploadChunk(
             @RequestParam("file") MultipartFile file,
             @RequestParam("chunkNumber") Integer chunkNumber,
-            @RequestParam("chunkSize") Long chunkSize,
-            @RequestParam("totalSize") Long totalSize,
+            @RequestParam("chunkSize") Integer chunkSize,
+            @RequestParam("totalSize") Integer totalSize,
             @RequestParam("identifier") String identifier,
             @RequestParam("filename") String filename,
             @RequestParam("totalChunks") Integer totalChunks,
